fixe a bug in the test harness.

+++ modules/engine/src/java/dalma/impl/Util.java
@@ -15,7 +15,9 @@ public class Util {
      *      if the operation fails.
      */
     public static void deleteContentsRecursive(File file) throws IOException {
-        for (File child : file.listFiles()) {
+        File[] files = file.listFiles();
+        if(files==null)     return; // non existent
+        for (File child : files) {
             if (child.isDirectory())
                 deleteContentsRecursive(child);
             if (!child.delete())
